Refactor lottery
